+++ myConvolution.h
@@ -11,8 +11,8 @@ enum FILTERING {
 	DERIVATIVE,
 };
 
-float calculation1(const Mat kernel, Mat pixels);
-float calculation2(const Mat kernel, Mat pixels);
+double calculation1(const Mat kernel, Mat pixels);
+double calculation2(const Mat kernel, Mat pixels);
 Mat myConvolution(const Mat src, Mat& kernel, FILTERING flag);
